Add script to download files from URLs
This commit adds a new script, `exec.bat`, which is used to download files from a list of URLs specified in the `urls.txt` file. The script checks if the `urls.txt` file exists and exits with an error if it doesn't. It then iterates through each URL in the file and uses `gsutil` to download the corresponding file to the `f:\mail2024` directory. Once all the files are downloaded, it prints a message indicating that the download is completed.

The `urls.txt` file is also added in this commit, which contains a list of URLs to download the files from.

This commit addresses the need to automate the process of downloading files from specific URLs.